requirements.txt
# --- Web Framework ---
sqlalchemy==2.0.44
fastapi==0.111.0
uvicorn[standard]==0.30.0
gunicorn==21.2.0

# --- HTTP Clients ---
httpx==0.27.0
requests==2.31.0

# --- Databases ---
motor==3.3.2
psycopg[binary]==3.2.1
redis==5.0.3

# --- Data validation & config ---
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# --- Optional (fuzzy matching) ---
fuzzywuzzy==0.18.0
rapidfuzz==3.14.0
celery==5.4.0
twilio
sendgrid==6.10.0
reportlab==4.2.2
asyncpg
scikit-learn










